requirements.txt
autopep8==1.6.0
awscli==1.22.87
botocore==1.24.32
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.3
colorclass==2.2.2
distlib==0.3.4
docopt==0.6.2
docutils==0.15.2
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
numpy
pywebio
packaging==21.3
pandas==1.3.4
pip-upgrader==1.4.15
pipenv==2022.1.8
pipreqs==0.4.11
platformdirs==2.4.1
psutil==5.9.0
psycopg2==2.9.2
pyasn1==0.4.8
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-gnupg==0.4.8
pytz==2021.3
pywebio==1.4.0
PyYAML==5.4.1
requests==2.27.1
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
tabulate==0.8.9
terminaltables==3.1.10
toml==0.10.2
tornado==6.1
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
virtualenv==20.13.1
virtualenv-clone==0.5.7
yarg==0.1.9
turtle==0.0.1
flask